refactor: VulnerabilityReportReconciler to WorkloadController
