+++ pkg/vulnerabilityreport/controller_test.go
@@ -1,0 +1,1 @@
+package vulnerabilityreport_test
